Use messageDisplayAction in messageDisplay windows
For some reason `browser_action` is not shown when a message
is opened in new window, always using `message_display_action`
is unnecessary. Using it instead of `browserAction` is confusing
since it may act on message list.

Thunderbird-91 has a problem with getting tabs from current
`messageDisplay` window, add a workaround.