Do not send to popup stale selection
If popup failed for some reason, do not send saved selection
to next popup.

Maybe it is reasonable to allow user to switch to previous
selection. It should be a workaround for issues like
absence of `browser_action` in `messageDisplay` windows,
and tricks to avoid such limitation that may have other bugs.